# JLBH: dynamic sleep duration when accounting for coordinated omission