@peter-lawrey 2017-05-30
The sleep of 1 ns is OS dependant as to how long it takes. On windows i
have seen it take 1 ms which would be a problem. I will fix this.


On 30 May 2017 09:06, "Alex Averbuch" <notifications@github.com> wrote:

> Previously, when accounting for CO, a sleep of rate - 1ns was *always*
> inserted between invocations -- even when the benchmark was running
> on/behind schedule.
> This resulted in accumulating delay, the longer a benchmark ran.
>
> To observe the behavior run this simple benchmark:
>
> public class NothingBenchmark implements JLBHTask
> {
>     private JLBH jlbh;
>
>     @Override
>     public void run( long startTimeNS )
>     {
>         jlbh.sample( System.nanoTime() - startTimeNS );
>     }
>
>     @Override
>     public void init( JLBH jlbh )
>     {
>         this.jlbh = jlbh;
>     }
> }
>
> With this config:
>
>         int iterations = 10; // 10, 50, 100
>         JLBHOptions lth = new JLBHOptions()
>                 .warmUpIterations( iterations )
>                 .iterations( iterations )
>                 .throughput( 5, SECONDS )
>                 .runs( 3 )
>                 .recordOSJitter( true )
>                 .accountForCoordinatedOmmission( true )
>                 .jlbhTask( new NothingBenchmark() );
>         new JLBH( lth ).start();
>
> Here is what I see on my laptop (*notice latencies rise as iterations
> rise*):
>
> --------------
> iterations = 10
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:         10747.90     13893.63     13893.63        16.33
> 90:         17301.50     21495.81     26738.69        26.67
> 99:         17301.50     21495.81     26738.69        26.67
> 99.9:       17301.50     21495.81     26738.69        26.67
> 99.99:      17301.50     21495.81     26738.69        26.67
> worst:      17301.50     21495.81     26738.69        26.67
> --------------
> iterations = 50
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:         40894.46     59768.83     53477.38        23.53
> 90:         81788.93    119537.66    106954.75        23.53
> 99:         94371.84    119537.66    119537.66        15.09
> 99.9:       94371.84    119537.66    119537.66        15.09
> 99.99:      94371.84    119537.66    119537.66        15.09
> worst:      94371.84    119537.66    119537.66        15.09
> --------------
> iterations = 100
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:        102760.45    102760.45    115343.36         7.55
> 90:        171966.46    197132.29    197132.29         8.89
> 99:        188743.68    205520.90    205520.90         5.59
> 99.9:      188743.68    213909.50    213909.50         8.16
> 99.99:     188743.68    213909.50    213909.50         8.16
> worst:     188743.68    213909.50    213909.50         8.16
> --------------
>
> And here's what I see with the changes in this commit:
>
> --------------
> iterations = 10
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:             2.75         2.37         2.75         9.76
> 90:            54.27        60.42        41.98        22.64
> 99:            54.27        60.42        41.98        22.64
> 99.9:          54.27        60.42        41.98        22.64
> 99.99:         54.27        60.42        41.98        22.64
> worst:         54.27        60.42        41.98        22.64
> --------------
> iterations = 50
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:             2.62         2.75         2.50         6.40
> 90:             8.96        31.23         6.53        71.61
> 99:            96.26        62.46        67.58        26.51
> 99.9:          96.26        62.46        67.58        26.51
> 99.99:         96.26        62.46        67.58        26.51
> worst:         96.26        62.46        67.58        26.51
> --------------
> iterations = 100
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:             2.62         2.37         1.95        18.67
> 90:            16.90         7.04         7.04        48.28
> 99:            88.06        75.78       116.74        26.49
> 99.9:         129.02        96.26       120.83        18.50
> 99.99:        129.02        96.26       120.83        18.50
> worst:        129.02        96.26       120.83        18.50
> --------------
>
> Although the 5ms value of LATE_MILLIS_THRESHOLD is somewhat arbitrary, it
> *was* selected based on experimentation.
> I experimented with several values of LATE_MILLIS_THRESHOLD (2ms, 5ms,
> 10ms) and compared against (a) original code and (b) with Jvm.pause(
> sleepMillis ) removed:
>
> iterations = 100
>
> --------------
> original
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:        102760.45    102760.45    115343.36         7.55
> 90:        171966.46    197132.29    197132.29         8.89
> 99:        188743.68    205520.90    205520.90         5.59
> 99.9:      188743.68    213909.50    213909.50         8.16
> 99.99:     188743.68    213909.50    213909.50         8.16
> worst:     188743.68    213909.50    213909.50         8.16
> --------------
> threshold = 2ms
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:          1277.95      1933.31      1802.24        25.48
> 90:          3211.26      3211.26      3342.34         2.65
> 99:          3473.41      3604.48      3866.62         7.02
> 99.9:        3735.55      3735.55      5373.95        22.62
> 99.99:       3735.55      3735.55      5373.95        22.62
> worst:       3735.55      3735.55      5373.95        22.62
> --------------
> threshold = 5ms
> --------------
> Percentile   run1         run2         run3      % Variation
> 50:             2.62         2.37         1.95        18.67
> 90:            16.90         7.04         7.04        48.28
> 99:            88.06        75.78       116.74        26.49
> 99.9:         129.02        96.26       120.83        18.50
> 99.99:        129.02        96.26       120.83        18.50
> worst:        129.02        96.26       120.83        18.50
> --------------
> threshold = 10ms
> --------------
> 50:             2.62         2.50         2.02        16.74
> 90:             5.50         5.25         6.53        13.99
> 99:            14.08        20.99        25.09        34.26
> 99.9:          18.94        33.79        54.27        55.42
> 99.99:         18.94        33.79        54.27        55.42
> worst:         18.94        33.79        54.27        55.42
> --------------
> Jvm.pause( sleepMillis ) removed
> --------------
> 50:             4.03         4.03         2.88        21.05
> 90:             9.47         7.55         7.55        14.49
> 99:            32.26        25.09        14.59        44.66
> 99.9:          41.98        26.11        18.94        44.78
> 99.99:         41.98        26.11        18.94        44.78
> worst:         41.98        26.11        18.94        44.78
> --------------
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/OpenHFT/Chronicle-Core/pull/30
> Commit Summary
>
>    - introduces dynamic sleep duration calculation, when accounting for
>    coordinated omission
>
> File Changes
>
>    - *M* src/main/java/net/openhft/chronicle/core/jlbh/JLBH.java
>    <https://github.com/OpenHFT/Chronicle-Core/pull/30/files#diff-0> (9)
>
> Patch Links:
>
>    - https://github.com/OpenHFT/Chronicle-Core/pull/30.patch
>    - https://github.com/OpenHFT/Chronicle-Core/pull/30.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/OpenHFT/Chronicle-Core/pull/30>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABBU8WvvLGgd3Uqds9pBuJ_oTA4ycfbZks5r-839gaJpZM4Np9AI>
> .
>