@alexaverbuch 2017-05-30
I tested on a Mac